+++ events/scripts/scripts.js
@@ -10,9 +10,8 @@
  * governing permissions and limitations under the License.
  */
 
-import { captureProfile } from '../utils/event-apis.js';
+import { lazyCaptureProfile } from '../utils/profile.js';
 import autoUpdateContent, { getNonProdData, setMetadata } from '../utils/content-update.js';
-import { getECCEnv } from '../utils/utils.js';
 import BlockMediator from '../deps/block-mediator.min.js';
 
 BlockMediator.set('imsProfile', null);
@@ -29,6 +28,26 @@ function getMetadata(name) {
   const attr = name && name.includes(':') ? 'property' : 'name';
   const meta = document.head.querySelector(`meta[${attr}="${name}"]`);
   return meta && meta.content;
+}
+
+function getECCEnv(miloConfig) {
+  const { env } = miloConfig;
+
+  if (env.name === 'prod') return 'prod';
+
+  if (env.name === 'stage') {
+    const { host, search } = window.location;
+    const usp = new URLSearchParams(search);
+    const eccEnv = usp.get('eccEnv');
+
+    if (eccEnv) return eccEnv;
+
+    if (host.startsWith('stage--') || host.startsWith('www.stage')) return 'stage';
+    if (host.startsWith('dev--') || host.startsWith('www.dev')) return 'dev';
+  }
+
+  // fallback to Milo env name
+  return env.name;
 }
 
 export function decorateArea(area = document) {
@@ -137,6 +156,6 @@ if ((window.eccEnv === 'stage' || window.eccEnv === 'dev') && !getMetadata('even
 (async function loadPage() {
   await loadLana({ clientId: 'events-milo' });
   await loadArea().then(() => {
-    captureProfile();
+    lazyCaptureProfile();
   });
 }());
